refactor(HttpResponse): MIME 타입을 지정하여 css 지원
