+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -8,9 +8,18 @@ public class HttpRequest {
         this.content = content;
     }
 
-    public String getTarget() {
+    public String getTarget() { //TODO 따로 빼야하나?
         String[] startLine = getStartLine().split(" ");
         return startLine[1];
+    }
+
+    public String getPath() {
+        String target = getTarget();
+        int i = target.lastIndexOf('?');
+        if (i == -1) {
+            return target; // login.html, login // login.css
+        }
+        return target.substring(0, i);
     }
 
     private String getStartLine() {
